Split tree dumper and checker into their own file; they deserve it
